stop centroding of non-DDA data from undoing lockmass correction
A cached centroided copy of the .raw was being created if vendor peakpicking was used.
This centroided raw was not lockmass corrected so lockmass correction did not happen if vendor peak picking was performed
Instead, read the non-IMS scan with a scanprocessor and use the samr scanprocessor to centroid if required.
vendor peakpicking does not currently work for Waters IMS data so this data has not been considered here