Fix error reading SONAR lockmass spectra
The assumption made in the code that the lockmass function does not have IMS (or quad separation for SONAR) data is incorrect. The LM function in HDMSE and SONAR data does have multiple scans per block
The SONAR_Short.raw tests were working before because the method isLockMassFunction(...) was incorrectly returning false for the lockmass function of the SONAR_short.raw injection.
Now isLockMassFunction(...) returns true for the lockmass function, the reader was trying to read LM data as none-IMS. This resulted in it trying to read numOfBlocks*blockPerScan spectra from the numOfBlocks non "IMS" representation of the
lockmass spectra causing the reader to fail with an "Invalid scan number" message